test: Split unittests into common, tensorflow and pytorch
