+++ test/common/test_datasets.py
@@ -1,0 +1,13 @@
+import pytest
+
+from doctr import datasets
+
+
+def test_visiondataset():
+    url = 'https://data.deepai.org/mnist.zip'
+    with pytest.raises(ValueError):
+        datasets.datasets.VisionDataset(url, download=False)
+
+    dataset = datasets.datasets.VisionDataset(url, download=True, extract_archive=True)
+    assert len(dataset) == 0
+    assert repr(dataset) == 'VisionDataset()'
