+++ test/common/test_utils_visualization.py
