+++ test/tensorflow/test_datasets.py
@@ -5,16 +5,6 @@ from doctr import datasets
 from doctr.transforms import Resize
 
 
-def test_visiondataset():
-    url = 'https://data.deepai.org/mnist.zip'
-    with pytest.raises(ValueError):
-        datasets.datasets.VisionDataset(url, download=False)
-
-    dataset = datasets.datasets.VisionDataset(url, download=True, extract_archive=True)
-    assert len(dataset) == 0
-    assert repr(dataset) == 'VisionDataset()'
-
-
 @pytest.mark.parametrize(
     "dataset_name, train, input_size, size, rotate",
     [
