+++ test/tensorflow/test_datasets_ocr.py
