+++ test/tensorflow/test_datasets_recognition.py
