+++ test/tensorflow/test_models_recognition.py
